Fixed NonClustered Index bug for Sql Server Temp Tables
Create nonclustered index for sql server temp tables included the
temporal_start_date column even when the column was not used in the
temporal query.  Resulted in an error being thrown by the database.  Now
check for the inclusion of the temporal start date before adding to
index.  Also added check for end date so that can also be included in
index.